Validate signed-scale Jolt physics transforms

+++ native/window_lifecycle_probe.h
@@ -248,9 +248,11 @@ inline bool probe_repeated_host_lifecycle() {
             telemetry.shutdown_functions_invoked != NativeApplication::MAX_SHUTDOWN_HOOKS ||
             telemetry.rejected_shutdown_registrations != 1 || telemetry.drained_callback_batches != 1 ||
             telemetry.peak_active_callbacks != 1) {
-            std::fprintf(stderr, "shutdown telemetry: peak_hooks=%u invoked=%u rejected=%u batches=%u peak_callbacks=%u\n",
-                telemetry.peak_shutdown_functions, telemetry.shutdown_functions_invoked,
-                telemetry.rejected_shutdown_registrations, telemetry.drained_callback_batches,
+            std::fprintf(stderr, "shutdown telemetry: peak_hooks=%zu invoked=%llu rejected=%llu batches=%llu peak_callbacks=%zu\n",
+                telemetry.peak_shutdown_functions,
+                static_cast<unsigned long long>(telemetry.shutdown_functions_invoked),
+                static_cast<unsigned long long>(telemetry.rejected_shutdown_registrations),
+                static_cast<unsigned long long>(telemetry.drained_callback_batches),
                 telemetry.peak_active_callbacks);
         }
         if (!check(host.window() == nullptr && host.close_requested() && reverse_order,
